chore: move reducers code to actions
